backend/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
python-dotenv>=1.0
google-generativeai>=0.8
pydantic>=2.5
pydantic-settings>=2.1
aiosqlite>=0.19
